Never load ActiveRecord models at load time.
This will only bring problems. For example, I didn't have the emailings table
but I couldn't run the migration to create it because it was trying to load the
table at load time and hence an exception was raised before the table was created,
leaving me in a blocked state.

[t: 30 minutes]